removed DocumentVerifier
