+++ identity-did/src/document/core_document.rs
@@ -5,6 +5,25 @@ use core::convert::TryInto as _;
 use core::fmt::Display;
 use core::fmt::Formatter;
 
+use identity_core::common::BitSet;
+use identity_core::common::Timestamp;
+use identity_core::crypto::merkle_key::Blake2b256;
+use identity_core::crypto::merkle_key::MerkleDigest;
+use identity_core::crypto::merkle_key::MerkleDigestTag;
+use identity_core::crypto::merkle_key::MerkleKey;
+use identity_core::crypto::merkle_key::MerkleSignature;
+use identity_core::crypto::merkle_key::MerkleSignatureTag;
+use identity_core::crypto::merkle_key::MerkleVerifier;
+use identity_core::crypto::merkle_key::Sha256;
+use identity_core::crypto::merkle_key::VerificationKey;
+use identity_core::crypto::Ed25519;
+use identity_core::crypto::JcsEd25519;
+use identity_core::crypto::ProofPurpose;
+use identity_core::crypto::Signature;
+use identity_core::crypto::TrySignature;
+use identity_core::crypto::Verifier;
+use identity_core::crypto::Verify;
+
 use serde::Serialize;
 
 use identity_core::common::Object;
@@ -20,11 +39,13 @@ use crate::error::Result;
 use crate::service::Service;
 use crate::utils::OrderedSet;
 use crate::verifiable::DocumentSigner;
-use crate::verifiable::DocumentVerifier;
+use crate::verifiable::Revocation;
+use crate::verifiable::VerifierOptions;
 use crate::verification::MethodQuery;
 use crate::verification::MethodRef;
 use crate::verification::MethodRelationship;
 use crate::verification::MethodScope;
+use crate::verification::MethodType;
 use crate::verification::MethodUriType;
 use crate::verification::TryMethod;
 use crate::verification::VerificationMethod;
@@ -565,6 +586,125 @@ impl<T, U, V> CoreDocument<T, U, V> {
   }
 }
 
+impl<T, U: Revocation, V> CoreDocument<T, U, V> {
+  /// Verifies the signature of the provided data.
+  ///
+  /// # Errors
+  ///
+  /// Fails if an unsupported verification method is used, data
+  /// serialization fails, or the verification operation fails.
+  pub fn verify_data<X>(&self, data: &X, options: &VerifierOptions) -> Result<()>
+  where
+    X: Serialize + TrySignature,
+  {
+    let signature: &Signature = data
+      .try_signature()
+      .map_err(|_| Error::InvalidSignature("missing signature"))?;
+
+    // Retrieve the method used to create the signature and check it has the required verification
+    // method relationship (purpose takes precedence over method_scope).
+    let purpose_scope = options.purpose.map(|purpose| match purpose {
+      ProofPurpose::AssertionMethod => MethodScope::assertion_method(),
+      ProofPurpose::Authentication => MethodScope::authentication(),
+    });
+    let method: &VerificationMethod<U> = match (purpose_scope, options.method_scope) {
+      (Some(purpose_scope), _) => self
+        .try_resolve_method_with_scope(signature, purpose_scope)
+        .map_err(|_| Error::InvalidSignature("method with purpose scope not found"))?,
+      (None, Some(scope)) => self
+        .try_resolve_method_with_scope(signature, scope)
+        .map_err(|_| Error::InvalidSignature("method with specified scope not found"))?,
+      (None, None) => self
+        .try_resolve_method(signature)
+        .map_err(|_| Error::InvalidSignature("method not found"))?,
+    };
+
+    // Check method type.
+    if let Some(ref method_types) = options.method_type {
+      if !method_types.is_empty() && !method_types.contains(&method.key_type) {
+        return Err(Error::InvalidSignature("invalid method type"));
+      }
+    }
+
+    // Check challenge.
+    if options.challenge.is_some() && options.challenge != signature.challenge {
+      return Err(Error::InvalidSignature("invalid challenge"));
+    }
+
+    // Check domain.
+    if options.domain.is_some() && options.domain != signature.domain {
+      return Err(Error::InvalidSignature("invalid domain"));
+    }
+
+    // Check purpose.
+    if options.purpose.is_some() && options.purpose != signature.purpose {
+      return Err(Error::InvalidSignature("invalid purpose"));
+    }
+
+    // Check expired.
+    if let Some(expires) = signature.expires {
+      if !options.allow_expired.unwrap_or(false) && Timestamp::now_utc() > expires {
+        return Err(Error::InvalidSignature("expired"));
+      }
+    }
+
+    // Check signature.
+    Self::do_verify(method, data)
+  }
+
+  /// Verifies the signature of the provided data matches the public key data from the given
+  /// verification method.
+  ///
+  /// # Errors
+  ///
+  /// Fails if an unsupported verification method is used, data
+  /// serialization fails, or the verification operation fails.
+  pub fn do_verify<X>(method: &VerificationMethod<U>, data: &X) -> Result<()>
+  where
+    X: Serialize + TrySignature,
+  {
+    let public_key: Vec<u8> = method.key_data().try_decode()?;
+
+    match method.key_type() {
+      MethodType::Ed25519VerificationKey2018 => {
+        JcsEd25519::<Ed25519>::verify_signature(data, &public_key)?;
+      }
+      MethodType::MerkleKeyCollection2021 => match MerkleKey::extract_tags(&public_key)? {
+        (MerkleSignatureTag::ED25519, MerkleDigestTag::SHA256) => {
+          merkle_key_verify::<X, Sha256, Ed25519, U>(data, method, &public_key)?;
+        }
+        (MerkleSignatureTag::ED25519, MerkleDigestTag::BLAKE2B_256) => {
+          merkle_key_verify::<X, Blake2b256, Ed25519, U>(data, method, &public_key)?;
+        }
+        (_, _) => {
+          return Err(Error::InvalidMethodType);
+        }
+      },
+    }
+
+    Ok(())
+  }
+}
+
+fn merkle_key_verify<X, D, S, U>(that: &X, method: &VerificationMethod<U>, data: &[u8]) -> Result<()>
+where
+  X: Serialize + TrySignature,
+  D: MerkleDigest,
+  S: MerkleSignature + Verify<Public = [u8]>,
+  U: Revocation,
+{
+  let revocation: Option<BitSet> = method.revocation()?;
+  let mut vkey: VerificationKey<'_> = VerificationKey::from_borrowed(data);
+
+  if let Some(revocation) = revocation.as_ref() {
+    vkey.set_revocation(revocation);
+  }
+
+  MerkleVerifier::<D, S>::verify_signature(that, &vkey)?;
+
+  Ok(())
+}
+
 // =============================================================================
 // Signature Extensions
 // =============================================================================
@@ -574,12 +714,6 @@ impl<T, U, V> CoreDocument<T, U, V> {
   /// signatures from verification methods in this DID Document.
   pub fn signer<'base>(&'base self, private: &'base PrivateKey) -> DocumentSigner<'base, '_, '_, T, U, V> {
     DocumentSigner::new(self, private)
-  }
-
-  /// Creates a new [`DocumentVerifier`] that can be used to verify signatures
-  /// created with this DID Document.
-  pub fn verifier(&self) -> DocumentVerifier<'_, T, U, V> {
-    DocumentVerifier::new(self)
   }
 }
