+++ identity-iota/src/document/iota_document.rs
@@ -29,7 +29,6 @@ use identity_did::document::CoreDocument;
 use identity_did::service::Service;
 use identity_did::utils::OrderedSet;
 use identity_did::verifiable::DocumentSigner;
-use identity_did::verifiable::DocumentVerifier;
 use identity_did::verifiable::VerifierOptions;
 use identity_did::verification::MethodQuery;
 use identity_did::verification::MethodRef;
@@ -518,15 +517,8 @@ impl IotaDocument {
   // Verification
   // ===========================================================================
 
-  /// Creates a new [`DocumentVerifier`] that can be used to verify digital signatures
-  /// created with this DID Document.
-  pub fn verifier(&self) -> DocumentVerifier<'_> {
-    self.document.verifier()
-  }
-
   /// Verifies the signature of the provided `data` was created using a verification method
-  /// in this DID Document. See [`IotaDocument::verifier`] for verifying signatures with
-  /// a builder pattern.
+  /// in this DID Document.
   ///
   /// # Errors
   ///
@@ -536,7 +528,7 @@ impl IotaDocument {
   where
     X: Serialize + TrySignature,
   {
-    self.verifier().verify(data, options).map_err(Into::into)
+    self.document.verify_data(data, options).map_err(Into::into)
   }
 
   /// Verifies that the signature on the DID document `signed` was generated by a valid method from
@@ -554,7 +546,7 @@ impl IotaDocument {
     let options = VerifierOptions::default()
       .method_scope(MethodScope::capability_invocation())
       .method_type(Self::UPDATE_METHOD_TYPES.to_vec());
-    signer.verifier().verify(signed, &options).map_err(Into::into)
+    signer.verify_data(signed, &options).map_err(Into::into)
   }
 
   /// Verifies whether `document` is a valid root DID document according to the IOTA DID method
@@ -626,7 +618,7 @@ impl IotaDocument {
     let options = VerifierOptions::default()
       .method_scope(MethodScope::capability_invocation())
       .method_type(Self::UPDATE_METHOD_TYPES.to_vec());
-    self.verifier().verify(diff, &options).map_err(Into::into)
+    self.verify_data(diff, &options).map_err(Into::into)
   }
 
   /// Verifies a [`DiffMessage`] signature and merges the changes into `self`.
